requirements.txt
pandas
snowflake-snowpark-python[pandas]
